More work on timeout events

+++ src/config/ParseKeySequence.h
@@ -56,7 +56,7 @@ private:
   Key read_key(It* it, const It end);
   void add_key_to_sequence(Key key, KeyState state);
   void add_key_to_buffer(Key key);
-  void add_timeout_event(KeyState state, uint64_t timeout);
+  void add_timeout_event(KeyState state, uint16_t timeout);
   bool remove_from_keys_not_up(Key key);
   void flush_key_buffer(bool up_immediately);
   void up_any_keys_not_up_yet();
